requirements.txt
django-user-accounts==1.3.1
pinax-stripe==3.2.1
pinax-webanalytics==2.0.4
pinax-eventlog==1.1.2
django-calendarium==1.3.4
social-auth-app-django==0.1.0
gunicorn==19.6.0
psycopg2==2.6.2

Django==1.9.8
pinax-theme-bootstrap==7.10.1

